requirements.txt
fusepy==3.0.1
redis
pylint
pep8

